+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/CodingKeysGenerator.kt
@@ -1,0 +1,8 @@
+package software.amazon.smithy.swift.codegen.integration
+
+import software.amazon.smithy.model.shapes.MemberShape
+import software.amazon.smithy.swift.codegen.SwiftWriter
+
+interface CodingKeysGenerator {
+    fun generateCodingKeysForMembers(ctx: ProtocolGenerator.GenerationContext, writer: SwiftWriter, members: List<MemberShape>)
+}
